+++ packages/kiarina-agi-data-builder/src/kiarina/agi/file_info_builder/_types/file_info_spec.py
@@ -30,6 +30,9 @@ class FileInfoSpec(TypedDict):
     start_time: NotRequired[float]
     end_time: NotRequired[float]
 
+    # video
+    analysis_fps: NotRequired[float]
+
     # pdf
     start_page: NotRequired[int]
     end_page: NotRequired[int]