+++ packages/kiarina-agi-data/src/kiarina/agi/file_bundle/_types/file_bundle_media_content_spec.py
@@ -10,3 +10,4 @@ class FileBundleMediaContentSpec(TypedDict):
     file_path: FileBundleFilePath
     mime_type: str
     visibility: NotRequired[FileBundleContentVisibility]
+    timestamp: NotRequired[float | None]